refactor(payment): dip 반영 아키텍처 수정
