+++ src/main/java/kr/hhplus/be/server/payment/application/PaymentService.java
@@ -2,7 +2,7 @@ package kr.hhplus.be.server.payment.application;
 
 import kr.hhplus.be.server.common.exception.PaymentException;
 import kr.hhplus.be.server.payment.domain.Payment;
-import kr.hhplus.be.server.payment.infrastructure.PaymentRepository;
+import kr.hhplus.be.server.payment.domain.repository.PaymentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
