+++ src/main/java/kr/hhplus/be/server/payment/domain/repository/PaymentRepository.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.payment.domain.repository;
+
+import kr.hhplus.be.server.payment.domain.Payment;
+
+public interface PaymentRepository {
+    Payment save(Payment payment);
+}
