--- src/main/java/kr/hhplus/be/server/payment/infrastructure/PaymentRepository.java
@@ -1,9 +1,0 @@
-package kr.hhplus.be.server.payment.infrastructure;
-
-import kr.hhplus.be.server.payment.domain.Payment;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
-}
