+++ src/main/java/kr/hhplus/be/server/payment/infrastructure/repository/PaymentJpaRepository.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.payment.infrastructure.repository;
+
+import kr.hhplus.be.server.payment.domain.Payment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PaymentJpaRepository extends JpaRepository<Payment, Long> {
+}
